Fixed a bug where Trees would merge incorrectly on case insensitive FSes
OSX's filesystem is case insensitive, but broccoli-merge-trees's cache
currently caches files like HISTORY.md and History.md as separate
entries, which means that it incorrectly handles conflicts in the filesystem.

A potential (better) fix could be to determine this based on whether the
FS is case sensitive or not, but that would result in different
behaviour on different platforms which could lead to very subtle bugs.

Not sure if this fix should land in broccoli main or only here (since it
really affects merging trees, this seems like a sensible place).